Fix: issue #2330 in Predication pass
Changes made in preorder of assignment and if statements, in order to fix the loss of assignment when dependencies exist.
New Predication class members:
Unsigned depNestingLevel used to track how deep in nesting do dependencies go.
AssignmentStatement pointer dependencyAssignement, added to indicate that dependencies exist, in which case no assignment is lost.
cstring dependantName used to track the assignemnt statements that are dependants.
vector of cstrings statNames used for storing assignement statements in preorder of assignment statements.
map isAssignemntPushed controls if the liveAssignemnts were already pushed on rv block.

Changes made:

In preorder of Assignment statements:
- Filling the liveAssignemnts map depending on the current dependencies nesting level and checking if the current statement is a dependant statement.
- elseStatementName cstring used for creating a new name for else branches in orderedNames.

In preorder of If statements:
- Checking if there were dependencies before pushing anything else on rv block (if dependnecyAssig$
- If the bool value of corresponding assignment is false, in the isAssignmentPushed map, the assignment should be pushed on rv block. If the bool value is true, it means that assignment has already been pushed when dependencies were handled.

In visitBranch of ExpressionReplacer:
- Fixing the assignment of condition variable from the right conditions vector element.