Fix to ingress_port assigned before ingress control is executed (#936)
The PSA spec says that ingress_port during ingress control should be
the same as it was sent into the ingress parser.  Without this fix,
ingress_port sent into the ingress control is always the original
input port on which the packet was received, which is incorrect for
recirculated packets.