Fix: Issue #2004 parser duplicated matches not optimized out
This fix contains solution for overlaping constant and range labels inside SelectExpression.
For example:
	3 : accept;
	0x3: reject; (This label will not exist in optimized code)
	5 .. 10: accept;
	1 .. 15: reject; (This range will be parsed in two ranges:
				1 .. 2: reject; (This part has 2 ovelapings- with label "3 : accept;" and "5 .. 10: accept;")
				11 .. 15: reject; (This part ovelaps only with label "5 .. 10: accept;"))
	6 .. 9: reject;	(This range will be removed because it's already asigned insidelabel "5 .. 10: accept;")

So, parsed code will look like this:
	3: accept;
	5 .. 10: accept;
	1 .. 2: reject;
	11 .. 15: reject;

Inside frontend.cpp new pass is added: RemoveDuplicatedSelectLabels which will remove all Constant and Range labels that are unreachable (constant that are duplicated, or inside some previous range, and duplicated range labels).
Inside p4test/midend.cpp new pass is added: OptimizeSelectLabels which will optimize select labels that are overlaping.
This code covers all posible overlapings between constants, ranges, and constants with ranges.

New SelectExpression will be created with new SelectCasses (that are from parsed label) and it will be forwarded to further processing.
If labels are not valid, this pass will not exclude or optimize them, they will be forwarded to further processing.

This code doesn't cover overlaping optimization of masks, Select Expression with structs and constants that are not Constant type.