Fix: Predication issue #2345
This commit fixes issue #2345.
The issue was the incorrect order of assignment statements with a function like:
	void dummy(inout Headers val) { }
	...
	action simple_action() {
        if (h.eth_hdr.eth_type == 1) {
            return;
        }
        h.eth_hdr.src_addr = 48w1;
        // this serves as a barrier
        dummy(h);
    }

Predication output now looks like the folowing code:
	@name("ingress.simple_action") action simple_action() {
		hasReturned = false;
		{
			bool cond;
			cond = h.eth_hdr.eth_type == 16w1;
			hasReturned = (cond ? true : hasReturned);
		}
		{
			bool cond_0;
			cond_0 = !hasReturned;
			h.eth_hdr.src_addr = (cond_0 ? 48w1 : h.eth_hdr.src_addr);
			val_eth_hdr = (cond_0 ? h.eth_hdr : val_eth_hdr);
			h.eth_hdr = (cond_0 ? val_eth_hdr : h.eth_hdr);
		}
	}

and before these changes were implemented these three lines were falsely aranged:

	{
		val_eth_hdr = (cond_0 ? h.eth_hdr : val_eth_hdr);
	}
	h.eth_hdr.src_addr = (cond_0 ? 48w1 : h.eth_hdr.src_addr);
	h.eth_hdr = (cond_0 ? val_eth_hdr : h.eth_hdr);

since h.eth_hdr.src_addr should be assigned to 1 bofore the function call.

The cause of the problem was the need to push a block with dependencies before liveAssignments. Which is now not an issue because all assignments are pushed on rv block from the same vector (liveAssigns), in preorder of if statements. In preorder of assignment statements, statements are pushed on liveAssigns in order they have on the input of predication pass.

The xor test was added since it first showed this problem. This test now also has a correct output. Also, some tests were added, which show adequate behaviour when mixing nested if else-es and dependencies.