Fixes #2613 predication issue
The removal of the condition on the line 192 leads to the correct predication output. The mentioned condition was not correct, it only tests if the travesalPath is false, which is true when there is any else branch in a p4 test, and the assignment in 'then' should be erased only when there is a statement with the same statementName in the else branch in p4 test.

Then-else handling is correctly managed in if block on line 201 (new line 196), so no other test outputs are changed by this commit.