Add INode::checkedTo for checked casts. (#2620)
Added a convenience method `INode::checkedTo` for checked casts. The
effect of `n->checkedTo<N>()` is similar to `&n->as<N>()`, except the
former produces an error message that is easier to debug. The former
also has clearer intent: the latter could be easily mistaken as being
equivalent to `n->to<N>()`.