Added test case with out-of-bounds index to header stack (#2666)
p4c already gives an error message if a header stack index is a
compile-time known value that is out of range, so there is no bug
here.  This is simply a test program that verifies that p4c does give
an error message for P4 programs that do this.