Print error code in logs for invalid runtime operations
This is especially useful when using P4Runtime, as PI does not preserve
the error code returned by bmv2 in case of invalid runtime operation
(e.g. when trying to insert a duplicate table entry). While most sanity
checks are performed at the P4Runtime server, in some cases (e.g. to
troubleshoot a bmv2 bug), it is useful to be able to inspect the bmv2
logs and see what the exact error is.

See https://github.com/p4lang/PI/issues/530