Update PSA test programs to always put parsed headers inside a struct (#2746)
* Update PSA test programs to always put parsed headers inside a struct
Many of these test programs were written so that a single Ethernet
header type was being passed as the parameter to top level parsers and
controls in the PSA architecture, rather than the more typical struct
of multiple headers.  This was not necessarily an error, but at least
the v1model architecture explicitly documents the user-defined headers
type H must be a struct containing header/stack/union types, but not
itself a header.  It would be best to avoid any special-case code in
the PSA back end to handle this unlikely use case.  I call it unlikely
because almost no one ever writes a program that only parses a single
header.

* Update expected output files for DPDK back end

* Updated expected outputs for one more test program