Remove dependency on Judy1 (#1002)
We have been experiencing issues with some builds of bmv2, with memory
errors in Judy when building it from source (it's possible that it's an
issue with how we build Judy and not with Judy itself...). Judy is very
complex and not actively maintained any more apparently, so it's
probably a good idea to remove the dependency on Judy altogether. In
this commit, we replace usage of Judy1 with alternative implementations:

* HandleMgr now uses a custom implementation, inspired from
  boost::dynamic_bitset.
* RandomAccessUIntSet is now implemented using
  boost::container::flat_set (a sorted vector which exposes a set
  interface). Note that this may introduce a dependency on a new Boost
  library (libboost-container-dev) for some distros.

Originally I was planning on supporting both the Judy-based
implementation and the alternative implementation (with a configure
switch). However I strongly believe that there is no significant
difference in performance, and it doesn't seem necessary to maintain
both versions.

We still need to remove the dependency on JudyL (LPM trie implementation
and BMI) which is likely to be a bit more difficult to do.

See #1001