Allow backend to skip input file check
Backends might want to implement custom --help-* options and in that
case they need a way to signal the driver that this option does not
require input source file to be set.

Since some sanity checks rely on the fact input has been provided,
we set the input to dummy.p4 so those checks still work.