Bazel rules for the graphs backend.
This adds
- a target (`:p4c_graphs`) to build the compiler binary with the graphs
  backend, and
- a build rule `p4_graphs` to invoke this backend and merge the graph
  files.

We need a new build rule (rather than specifying a different binary for
`_p4c` in the `p4_library` rule) because the graphs backend does
not use the normal output file arguments and uses a separate argument to
specify the output *directory* to write the graphs in. The names of the
output files depend on the content of the input p4 file, and blaze
outputs cannot be directories (they can be file groups but we do not
know the exact file names without effectively running the p4 compiler).
So, we cannot directly use the output files in a blaze target. Instead,
this build rule concatenates the output graphs into a single GraphViz
file. The output file can later be processed by `gvpack` and `dot` to
produce a single large graph (the original graphs are embedded as
subgraphs). It can also be parsed by other tools to extract specific
graphs.