Use egress_spec instead of egress_port in parser inlining tests
According to:
https://github.com/p4lang/behavioral-model/blob/main/docs/simple_switch.md

egress_spec - Can be assigned a value in ingress code to control which
output port a packet will go to.
egress_port - Only intended to be accessed during egress processing,
read only. The output port this packet is destined to.

So this fixes a mistake made previously when I thought that
egress_port determines the port used for egress.