Add compiler option for disabling optimization of parser inlining
Optimization of subparser inlining transforms the parser in a way,
which may cause troubles for some backends in some special cases by
adding more constraints and requirements for the target which may
cause compilation error because some constraints of the target are
not satisfied, while compilation without the optimization may pass.

Because of that new command line option is introduced which allows
to disable the optimization.