Use mmap instead of sbrk (#2925)
* Use mmap instead of sbrk

For memory that needs to be allocated before GC is initialized.
sbrk is not portable and some standard library implementations (musl) do
not actually implement it at all.
Instead we use mmap, which should be more portable. We can now also
support releasing that memory when it is not needed anymore, instead of
leaking it.