dpdk: support of PNA extern send_to_port
send_to_port() extern for PNA architecture has been added recently,
but instructions were not generated properly.
This commit fixes the generated instructions.
Metadata field used as operand for 'tx' instruction is added to
the main metadata structure.
The name of the field set when send_to_port extern is used is unified
with the name of the field used as operand for 'tx' instruction.