Fix undefined behavior in GlobalCopyPropagation pass
The code was calling a method on a NULL pointer.
While this will work most of the time with most compilers, this is still
undefined behavior. As a result, when compiling with O3 the compiler may
assume that the method is never called with a NULL pointer to optimize
code.
In our case the compiler can assume that to<T> is not called on a NULL
pointer and therefore that dynamic_cast is not called with a NULL
pointer. This can lead to the NULL pointer check in dynamic_cast
([example](https://github.com/gcc-mirror/gcc/blob/29a1af24eface3620e348be9429e7c2e872accbc/libstdc%2B%2B-v3/libsupc%2B%2B/dyncast.cc#L50-L51))
being eliminated and we end up with a runtime crash.

Fixes #2944