dpdk: change varbit test for PSA to work around issue in ParsersUnroll
We were using field of invalid header in the argument for extract
method (using lookahead and then assigning the value to a header does
not make that header valid), which is causing in ParsersUnroll pass
to drop some statements from that state.
As the header is invalid when we are accessing its field, the behavior
is undefined, so this commit changes the test to avoid having
undefined behavior.