dpdk: change varbit tests for psa and pna arch
PNA and PSA tests for varbit did not allow to distinguish the cases
when the header with varbit field was parsed and when it was not
and was just considered as payload.

Examples of packets and expected outputs are also in a new stf file.
Using this stf file, the test for PSA arch can be also run on bmv2
target.

When the tests are compiled for dpdk target, following warning is
displayed:
warning:  pkt.extract is ignored. Error: Reading field from invalid header
        pkt.extract(hdr.ipv4_option_timestamp, (bit<32>)(hdr.ipv4_option.len) * 8 - 16);

The origin of this warning is in ParsersUnroll pass during symbolic
evaluation of lookahead method.