dpdk: change varbit tests for psa and pna arch (#2993)
PNA and PSA tests for varbit did not allow to distinguish the cases
when the header with varbit field was parsed and when it was not
and was just considered as payload.

Examples of packets and expected outputs are also in a new stf file.
Using this stf file, the test for PSA arch can be also run on bmv2
target.

When the tests are compiled for dpdk target, following warning is
displayed when a lookahead with header is used:
warning:  pkt.extract is ignored. Error: Reading field from invalid header
        pkt.extract(hdr.ipv4_option_timestamp, (bit<32>)(hdr.ipv4_option.len) * 8 - 16);

The origin of this warning is in ParsersUnroll pass during symbolic
evaluation of lookahead method.

When lookahead with header is used, the validation of programs also fails,
therefore instead of lookahead with header we use combination of lookahead
with bit<16> and slicing.