Inline Alloc{1,2}D storage
This changes `Alloc{1,2}Dbase` to be a non-owning view into one- or two-
dimensional array, and `Alloc{1,2}D` to hold the actual data inline
instead of letting AllocXDbase to own the elements in manually managed
heap-allocated memory.

The change improves mainly performance of the class. It also
simplifies the class a bit as a bonus.