Update the p4c Docker image periodically
Other p4lang Docker images are now updated periodically. We do the same
thing for p4c to pick up updates to the base images even when the p4c
repo itself is not updated.
We update the image every 4 hours. To avoid using using too much compute
in Github actions, we update the workflow definition to 1) use native
Docker actions and 2) leverage the Github actions cache. When the base
image (p4lang/behavioral-model) has not been updated and there are no
changes to the p4c code, the workflow will complete in less than a
minute.