dpdk: use header instead of metadata for lookahead in varbit tests
DPDK pipeline library supports lookahead instruction only with header
not with metadata.
Local variable in parser is internally transformed to a main metadata
field which is currently not allowed as an operand for a lookahead
instruction in DPDK pipeline library.
Therefore we currently need to use a header which is part of main
headers structure.

Also we should handle somehow the case when lookahead is used with
local variable, because that is definitely valid scenario, but
at the moment the generated spec file can not be processed by DPDK
backend.

These tests are intended to test varbit feature, not lookahead, so
this is fixing only the tests so that they can be used to test
varbit with DPDK pipeline library.

Lookahead with local variables can be fixed separately.