Release 1.2.2.1
The current versioning scheme of the P4 compiler corresponds to the P4
language specification. In particular, the current compiler version
corresponds to v1.2.2 of P4_16 Language Specification.

This patch updates the P4 compiler version to the following format

	<major>.<minor>.<patchlevel>.<releasenumber>

Where <releasenumber> increments with each release of the compiler.
The increment in version number allows to distribute updates for the
P4 compiler using package management systems such as
DPKG (Debian/Ubuntu) and RPM (Fedora/CentOS).

This change was discussed at the P4 open source developer days on
November 23, 2021.

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>