ordered_(set|map): Increased exception safety via the new `AutoEraseGuard`.
Regarding a sequence of _related_ insertions such as
```
container1.insert (...); container2.insert (...);
```
:

In order to achieve at least the basic exception safety guarantee,
if the second insertion fails, before propagating the exception,
we wish to try to undo the first insertion.
This is a newly introduced change.

Additionally, to ease debugging, if no `catch` clause for the exception
has existed (up the call stack), we wish that the core file
captures the exact moment of the first failure.
This has been the existing behaviour, and we strive to keep it.
It would have been badly changed by code such as:

```
container1.insert (...);
try { container2.insert (...); } catch (...) { container1.erase (...); throw; }
```