backends/ebpf: Output an error if a header is not byte-aligned (#4176).
Although P4 allows headers to be an arbitrary number of bits, the
eBPF backend does not support it properly. Specifically, the code
in compileExtractField() assumes that the header starts at a byte
boundary. In any case, there is no realistic use case for headers
which aren't a whole number of bytes. Output an error rather than
fail silently. The BMv2 backend has a similar limitation already.