Add validation of constant enum values fitting
According to P4 spec v1.2.4 section 7.2.1 Enumeration types, the serialized enums should raise error if a value is not representable in the underlying type.  This was not the case for p4c. The implementation is complicated a little bit by the fact it has to run before type checking (as that cuts the constants, emitting just a warning). Therefore, the implementation has to resolve type names manually.