Add validation of constant enum values fitting (#4287)
According to P4 spec v1.2.4 section 7.2.1 Enumeration types, the serialized enums should raise error if a value is not representable in the underlying type. This was not the case for p4c. I opted to add this check to type checking as it cannot run after it (as type checking cuts the constants, emitting just a warning), but running it before is more complicated as the type would need to be resolved manually. Therefore, the implementation has to resolve type names manually.

User-visble changes

    Compiler now rejects enumeration constants in serialized enums that are not representable in the underlying type. This now matches with the P4 spec.
    If for some reason a constant that does not fit is desired, it can be defined with an explicit cast.