Fix build with system Protobuf
This sets the flag `protobuf_MODULE_COMPATIBLE` in the Protobuf cmake
file, so the Protobuf files expose some variables required by the build,
such as `Protobuf_PROTOC_EXECUTABLE`.

Furthermore, `Protobuf_LIBRARY` is defined manually as
`protobuf::libprotobuf`, because it is only holds the path to the
Protobuf library, not the abseil library Protobuf depends on since
version 22.

Fix #4316