Remove XDP code generated for TC
Remove XDP section that is always generated for the TC backend.
This XDP code generated for TC was responsible for setting the
skb->protocol field which is not writable in TC. We have a found better
way to do that with an extern, hence this XDP section isn't needed
anymore.