backends/ebpf: Track header offset in bytes rather than bits. (#4327)
* backends/ebpf: Track header start offset in bytes rather than bits.

* backends/ebpf: Remove most remaining uses of bits offset var.

Uses still remain in ubpf and tc backends. It would be easy to move the
variable definition to the uBPFProgram class, however TC uses eBPFProgram
directly (without inheritance).

ubpf could be changed to use a byte-level header var as was done in eBPF,
however it supports bit-level advance() calls.

* testdata: Update p4tc golden outputs following offset tracking changes.

This patch also sets the executable bit on all .template files. This is
done by the TC backend when they are created, however for some reason
never made it to the reference outputs, and the checker script doesn't
look at permissions.