Make sure drop starts out false
We found scenarios where the skb->cb is not zeroed out before the bpf
program is called. Since compiler_meta__ = skb->cb, there is an issue in
some specific scenarios. For example, since the bpf code looks at
compiler_meta__->drop to tell whether the packet should be dropped or not,
if by any chance compiler_meta__->drop is initialised to true, the
program will mistakenly drop the packet. To solve this, we are
initialising compiler_meta__->drop to false before doing any
processing on compiler_meta__->drop.