Add back in heading for P4Smith README file (#4756)
* Add back in heading for P4Smith README file
This makes the auto-generated doxygen index easier to distinguish this
files from others, instead of merely being called "README" there.

Signed-off-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>

* Trying out a thing where the P4 logo is put before the CI badge
Maybe not as nice of a visual look as what was there before, but
seems worth looking at to see if it is desirable.

Signed-off-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>

* Removing logo
since previous attempt to keep it messed up the CI badge when viewing
Github flavored Markdown on github.com

Signed-off-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>

* Also remove extra heading for P4Testgen

Signed-off-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>

---------

Signed-off-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>
Signed-off-by: Adarsh <Adarshbunny293@gmail.com>