Reduce extern parameters size to 64 bytes
eBPF has, currently, a stack limit of 512 bytes.
In a scenario where we have declared one instance of
struct p4tc_ext_bpf_params, another of struct p4tc_ext_bpf_val and one of
struct p4tc_table_entry_act_bpf we have already 420 bytes of stack usage.
With a few extra stack variables it's possible to see how the limit will
be easily exceeded. To avoid exceeding this limit so trivially, reduce the
size of the parameters in the extern structs. As of now, we don't have a
use case that would require more than 64 bytes.

Signed-off-by: Victor Nogueira <victor@mojatatu.com>