Handle pna_main_parser_input_metadata_t fields in the tc backend's parser (#4955)
* Handle pna_main_parser_input_metadata_t fields in the tc backend's parser

As of now, referencing a field of pna_main_parser_input_metadata_t will
cause the it to be generated in eBPF without any translation.

For example:

parser Ingress_Parser(
        packet_in pkt,
        out   my_ingress_headers_t  hdr,
        inout my_ingress_metadata_t meta,
        in    pna_main_parser_input_metadata_t istd)
{
    state mystate {
        ...
        meta->ingress_port = istd.input_port;
        ...
    }
    ...
}

Will generate something like:

mystate: {
    ...
    meta->ingress_port = istd.input_port;
    ...
}

This will cause a compilation error because istd is not defined in the ebpf
code. This commit fixes this by either translating the fields to something
ebpf understands or simply throwing an error if the field is unsupported.

Signed-off-by: Victor Nogueira <victor@mojatatu.com>

* Add example to illustrate accessing parser metadata

Signed-off-by: Victor Nogueira <victor@mojatatu.com>

---------

Signed-off-by: Victor Nogueira <victor@mojatatu.com>