Fix action parameter declaration in p4tc_filter_fields in tc_may_override cases
In cases where the tc_may_override annotation is associate with an
action that has a parameter with a tc_type annotation, the compiler
is emitting the parameters declaration in p4tc_filter_fields incorrectly.
For example in tc_may_override_example_04:

struct p4tc_filter_fields {
    __u32 pipeid;
    __u32 handle;
    __u32 classid;
    __u32 chain;
    __u32 blockid;
    __be16 proto;
    __u16 prio;
    ipv4 ipv4_tbl_1_next_hop_ipv4addr;
};

"ipv4" is not a valid C type, so this won't compile

Also, in the case of bit types that are not equivalent to a standard C
type (u8, u16, u32, u64), the compiler is emitting the type incorrectly.
Let's say, in the previous example, that the type for ipv4addr was bit<33>,
the compiler would've emitted the following:

struct p4tc_filter_fields {
    __u32 pipeid;
    __u32 handle;
    __u32 classid;
    __u32 chain;
    __u32 blockid;
    __be16 proto;
    __u16 prio;
    __u33 ipv4_tbl_1_next_hop_ipv4addr;
};

This commit fixes this by converting tc_type annotated action parameters
to their corresponding C type and converts any non C standard bit type
into a u8 array rounded up to the nearest byte

Signed-off-by: Victor Nogueira <victor@mojatatu.com>