Ignore typeless structs in unstructured annotations
This pass expects struct expressions in unstructured annotations to have
explicit types. Using { } in unstructured annotation is actually allowed
by the language spec, it does not make sense for the frontend to prevent
this (in fact any sequence of tokens can be used as long as parentheses
are balanced). nvp4c will need to use untyped struct expressions in this
context to allow the user to provide values for the fields of headers.

Signed-off-by: Chris Dodd <cdodd@nvidia.com>