Remove unused files from testdata
All of the files removed by this commit appear never to be used during
p4c test runs.  Some details about each removed file and why it is
still in the repository are given below.

+ p4_14_samples_outputs/wide_action2-midend.p4

The other files with names beginning "wide_action2" were removed with
this commit in 2019.  It appears that the one remaining one was not
removed then, as it should have been.

commit ae6fb228a0d307466f482affd2b47fbd079f3c9e
Author: Mihai Budiu <mbudiu@vmware.com>
Date:   Thu Apr 25 16:46:31 2019 -0700

+ p4_16_dpdk_errors_outputs/pna-dpdk-large-constants.p4-error
+ p4_16_dpdk_errors_outputs/pna-dpdk-large-constants.p4.bfrt.json

Other files with a similar name were modifed, or moved from directory
p4_16_dpdk_errors_outputs to directory p4_16_samples_output, with this
commit.  The two files above appear to have been forgotten to be
removed.

commit 4005380cbaecb95de711ce50c41d6f6c24e97c46 (HEAD -> main, origin/main, origin/
HEAD)
Author: maheswari-s <maheswari.subramanian@intel.com>
Date:   Mon Dec 23 22:41:14 2024 +0530

+ p4_16_errors/constructor3_e.s

It was first added in this commit:

commit 6b9da239267529aeea6946cb0aa984c4487b07d9
Author: Mihai Budiu <mbudiu@vmware.com>
Date:   Mon Nov 23 11:58:14 2020 -0800

which appears to have been a mistake.

+ p4tc_samples_outputs/default_action_example_parser.o

This is an ELF format binary file, which appears to have been
committed by accident in this commit.

commit 900572921c0c1626fe42eee56d98d2c87c8d58ac
Author: Komal Jain <komal.jain@intel.com>
Date:   Wed Nov 29 16:34:20 2023 +0530

+ p4_16_errors_outputs/issue1777-bmv2.p4.entries.txt
+ p4_16_errors_outputs/issue1803_same_table_name.p4.entries.txt
+ p4_16_errors_outputs/issue-2123_e.p4.entries.txt
+ p4_16_errors_outputs/issue2283_2-bmv2.p4.entries.txt
+ p4_16_errors_outputs/issue3299.p4.entries.txt
+ p4_16_errors_outputs/issue532.p4.entries.txt
+ p4_16_errors_outputs/table-entries-lpm-2.p4.entries.txt
+ p4_16_errors_outputs/table-entries-optional-2-bmv2.p4.entries.txt
+ p4_16_samples_outputs/fabric_20190420/fabric.p4.entries.txt
+ p4_16_samples_outputs/omec/up4.p4.entries.txt
+ p4_16_samples_outputs/pins/pins_fabric.p4.entries.txt
+ p4_16_samples_outputs/pins/pins_middleblock.p4.entries.txt
+ p4_16_samples_outputs/pins/pins_wbb.p4.entries.txt

All of the files above have another file in the repo with a file name
suffix of `.txtpb`, and the commit below that added many other
`.txtpb` files and removed the corresponding `.txt` files neglected to
remove the files above.

commit 5082fe343f9f13a514c92918b93cd01578584953
Author: Fabian Ruffy <5960321+fruffy@users.noreply.github.com>
Date:   Wed Jan 24 18:31:00 2024 +0100

Signed-off-by: Andy Fingerhut <andy_fingerhut@alum.wustl.edu>