Add redirect_l2 test case
The redirect_l2 P4 program redirects an IPv4 packet to a specific port and
changes the packet's source and destination mac addresses. It does that
whenever the source IP address of the parsed packet matches a table
entry. The entry's associated action will specify the new source and
destination mac addresses along with the port where the program should
redirect to.

In this test case, we add a table entry which redirects packets with
source IP address 10.0.0.1 to port 1 and changes the mac source and
destination addresses, respectively, to 00:01:02:03:04:05 and
06:07:08:09:11:12. The test case will send an IPv4 packet with source
address 10.0.0.1 to port 0 and check whether the redirected IPv4 packet
was sent to port 1

Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-authored-by: Jamal Hadi Salim <jhs@mojatatu.com>