Add p4tc IP-in-IP tunnel test cases
Add 2 tunnel test cases using ipip.

They both use the same P4 program which implements IP-in-IP packet
encapsulation and decapsulation (ipip.p4).

The ipip program parses plain IPv4 as well as IP-in-IP packets. Any
other packets are rejected (dropped).

A lookup, using as key the port/netdev the packet arrived on, is applied
on table fwd_table. A match will either result in action set_ipip or
set_nh. If an entry has no action or no entry is found, then by
default drop action is assumed, meaning the packet will be dropped.

The set_ipip action facilitates pushing an IP-in-IP header and defining
the egress port to send on. The set_nh action sets a programmed
destination ethernet mac address and selects the egress port to send on.

From the above it should be noted that in our simple example - since our
lookup key is based on the port on which the packet ingressed - to make
it meaningful/observable we would need at least two entries: For one
port which received IP-in-IP packets to program it with an action
set_nh and for the other port that receives ip packets to program it
with set_nh. Then we send IP-in-IP packets on one port and plain IP
packets on another. Note, it is not wrong to send packets on a single
port programmed with set_ipip action, but it will be less "observable"
because we wont see the changed mac address that is provided by the
set_nh action.

Our test cases do the following:

ipip_encap:
- Sends a plain ip packet to port0
- Packet triggers call to set_ipip action
- Adds an outer IP header with the source and destination
  addresses assigned by the set_ipip action (2.2.2.2 and 2.4.4.8)
- Redirects to port specified in set_ipip action (port1)

ipip_decap:
- Sends an ipip packet to port1
- Packet triggers call to set_nh action
- Removes the outer IP header
- Overwrites destination mac address with value specified in set_nh
  action (66:33:34:35:46:01)
- Redirects to port specified in set_nh action (port0)

Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-authored-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-authored-by: Pedro Tammela <pctammela@mojatatu.com>